requirements.txt
alembic==1.8.0
blinker==1.8.2
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
itsdangerous==2.0.1
Jinja2==3.1.5
Mako==1.1.5
MarkupSafe==2.1.2
PyMySQL==1.0.2
python-dotenv==0.21.0
SQLAlchemy==1.4.41
typing_extensions==4.4.0
Werkzeug==2.2.3
requests==2.28.1
flask-bcrypt==1.0.1 # 비밀번호 해싱 및 검증
pyjwt==2.7.0 # JWT 생성 및 검증
redis
pytz